feat: call printerServer ip registration on first venueId receival
